+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - the one column mode is now turned off by default, it was
   accidentally turned on by default
   (for [Issue #281](https://github.com/pbek/QOwnNotes/issues/281))
+	- the note edit pane will now be automatically turned on
+	  when entering the one column mode
 
 ## 16.08.6
 - there now is a **one column minimal mode**, like NVAlt has